Exploratory analysis of coffee dataset
This file contains the numerical summaries and exploratory plots of the coffee quality dataset. The important point to note here is that there were many significant outliers in the model. It should also be observed that since the sample size is small so removing them all could be risky. To deal with this only those outliers have been removed which were significantly impacting the boxplots